requirements.txt
torch_geometric
pyalex
requests
numpy
tqdm
pandas
sentence-transformers